[rbd] allow calculating real size of an image
The image size is reported by the virtual size of the rbd image,
thus not reflecting the real amount occupied on the disk.

Using diff_iterate, we're able to iterate over the image blocks
and calculate the real image size.

This can be enabled on purpose, by setting rbd_size_by_diff=True